ROE-2216 Start new data retieval service
